Refactoring spatial spotdata utilities into spatial.py
