+++ lib/improver/tests/spotdata/spotdata/test_common_functions.py
@@ -36,15 +36,12 @@ import warnings
 from datetime import datetime as dt
 
 import cf_units
-from iris.coords import (DimCoord,
-                         AuxCoord)
-from iris import coord_systems
+from iris.coords import DimCoord
 from iris.coord_systems import GeogCS
 from iris import Constraint
 from iris.cube import Cube, CubeList
 from iris.tests import IrisTest
 from iris.time import PartialDateTime
-import cartopy.crs as ccrs
 import numpy as np
 
 from improver.spotdata.common_functions import (
@@ -380,20 +377,6 @@ class Test_node_edge_check(Test_common_functions):
         self.assertArrayEqual(expected, result)
 
 
-class Test_get_nearest_coords(Test_common_functions):
-    """Test wrapper for iris.cube.Cube.nearest_neighbour_index."""
-
-    def test_nearest_coords(self):
-        """Test correct indices are returned."""
-        plugin = get_nearest_coords
-        longitude = 80
-        latitude = -25
-        expected = (4, 8)
-        result = plugin(self.cube, latitude, longitude,
-                        'latitude', 'longitude')
-        self.assertEqual(expected, result)
-
-
 class Test_index_of_minimum_difference(Test_common_functions):
     """
     Test ability to identify the index of a minimum value in an array,
@@ -504,61 +487,31 @@ class Test_apply_bias(Test_common_functions):
         self.assertArrayEqual(expected, result)
 
 
-class Test_xy_determine(Test_common_functions):
-    """Test function that tests projections used in diagnostic cubes."""
-
-    def test_projection_test(self):
-        """Test identification of non-lat/lon projections."""
-        src_crs = ccrs.PlateCarree()
-        trg_crs = ccrs.LambertConformal(central_longitude=50,
-                                        central_latitude=10)
-        trg_crs_iris = coord_systems.LambertConformal(
-            central_lon=50, central_lat=10)
-        lons = self.cube.coord('longitude').points
-        lats = self.cube.coord('latitude').points
-        x, y = [], []
-        for lon, lat in zip(lons, lats):
-            x_trg, y_trg = trg_crs.transform_point(lon, lat, src_crs)
-            x.append(x_trg)
-            y.append(y_trg)
-
-        new_x = AuxCoord(x, standard_name='projection_x_coordinate',
-                         units='m', coord_system=trg_crs_iris)
-        new_y = AuxCoord(y, standard_name='projection_y_coordinate',
-                         units='m', coord_system=trg_crs_iris)
-
-        cube = Cube(self.cube.data,
-                    long_name="air_temperature",
-                    dim_coords_and_dims=[(self.cube.coord('time'), 0)],
-                    aux_coords_and_dims=[(new_y, 1), (new_x, 2)],
-                    units="K")
-
-        plugin = xy_determine
-        expected = trg_crs
-        result = plugin(cube)
-        self.assertEqual(expected, result)
-
-
-class Test_xy_transform(Test_common_functions):
+class Test_extract_cube_at_time(Test_common_functions):
     """
-    Test function that transforms the lookup latitude and longitude into the
-    projection used in a diagnostic cube.
+    Test wrapper for iris cube extraction at desired times.
 
     """
-    def test_projection_transform(self):
-        """
-        Test transformation of lookup coordinates to the projections in
-        which the diagnostic is provided.
 
-        """
-        trg_crs = ccrs.LambertConformal(central_longitude=50,
-                                        central_latitude=10)
+    def test_valid_time(self):
+        """Case for a time that is available within the diagnostic cube."""
+        plugin = extract_cube_at_time
+        cubes = CubeList([self.cube])
+        result = plugin(cubes, self.time_dt, self.time_extract)
+        self.assertIsInstance(result, Cube)
 
-        plugin = xy_transform
-        expected_x, expected_y = 0., 0.
-        result_x, result_y = plugin(trg_crs, 10, 50)
-        self.assertAlmostEqual(expected_x, result_x)
-        self.assertAlmostEqual(expected_y, result_y)
+    def test_invalid_time(self):
+        """Case for a time that is unavailable within the diagnostic cube."""
+        plugin = extract_cube_at_time
+        time_dt = dt(2017, 2, 18, 6, 0)
+        time_extract = Constraint(time=PartialDateTime(
+            time_dt.year, time_dt.month, time_dt.day, time_dt.hour))
+        cubes = CubeList([self.cube])
+        with warnings.catch_warnings(record=True) as w_messages:
+            plugin(cubes, time_dt, time_extract)
+            assert len(w_messages) == 1
+            assert issubclass(w_messages[0].category, UserWarning)
+            assert "Forecast time" in str(w_messages[0])
 
 
 class Test_extract_ad_at_time(Test_common_functions):
